requirements.txt
PyQt5    ===== 5.15.3
numpy    ===== 1.22.1
pyserial ===== 3.5
pyzbar   ===== 0.1.8